requirements.txt
streamlit
torch
torchvision
timm
numpy
opencv-python-headless
pillow
streamlit-drawable-canvas
